requirements.txt
Flask==2.0.1
Pillow==9.0.0
torch==2.2.0
torchvision==0.17.0
Werkzeug==2.0.1
